When the user enters text in the logchefql or SQL Monaco editor using the {{variable_name}} format, an input field labeled with variable_name appears at the top of the editor. The value entered by the user in this input field is then reflected in the query when executed.
When the user clicks on the input field, a settings drawer appears on the right, allowing the user to select the input type (text, number, or date).

The user can save the query with these dynamic variables in place, and when the saved query is later loaded, the last settings for each variable are also restored.